feat: modelo e marca

+++ Back/DTO's/ReservaDTO.cs
@@ -2,7 +2,7 @@
 {
     public class ReservaDTO
     {
-        public long CarroId { get; set; }
+        public int CarroId { get; set; }
         public int ClienteId { get; set; }
 
         //public DateOnly? dataDeReserva { get; set; } = DateOnly.FromDateTime(DateTime.Now);
